requirements.txt
streamlit
pandas
requests
google-genai
openpyxl
urllib3